Don’t match search w letters in mid of ingredients
This contributes to https://www.pivotaltracker.com/story/show/182936494
which was mainly done in f5102b84babc11855c5e2f98115d2fcfa0333ffc.

This commit acknowledges the note at the bottom of the story which
states:

Results are filtered as the ingredient is typed; eg if `O` is typed to
begin searching for the ingredient "Orange Juice", recipes requiring the
ingredient "Orange" or "Orange Juice" will be shown in the list of
applicable drink recipes, however recipes requiring "Bourbon" will not
be shown (as it might be thought since there is also an `o` in bourbon).