Add alternate ingredients for all ingredient types
Also, add a prefix to measured and numbered ingredients (so now every
ingredient has a prefix - the other type is portionless which has both
prefix and suffix).

We think maybe we should remove suffix from portionless ingredients.
They’re not really suffixes (eg, lime wheel, orange twist).

Using them is kind of a shortcut to keep the ingredient and ingredient
alias lists shorter for faster filtering (it’s super fast atm anyway).

This commit also unifies how ingredients are displayed in both the
ingredient list and in the drink making instructions. We can pull those
two paths apart again later if we want to. For now, we’ve been updating
it in two places every time we make a change and both places had been
slopily maintained and were a clear source of “I don’t want to touch
those” for developers so we unified the code and tried to make it nicer
to touch.